app/requirements.txt
fastapi==0.89.1
tortoise-orm==0.19.2
aerich==0.7.1
asyncpg==0.27.0
pydantic==1.10.4
uvicorn==0.20.0
celery==5.2.7
requests
